+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## v0.3.9 (2024-05-09)
+* Added Cache Purge task to BuildManifest
+* Fixed realpath in ensureCliSource
+
 ## v0.3.8 (2024-05-09)
 * Added exit to ensureCliSource
 * Added --fabric-source flag to update task